app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["reloverevolt <savitskiy.revenway@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "^3.14.0"
Django = "4.2.10"
psycopg2-binary = "^2.9.9"
django-extensions = "^3.2.3"
ipython = "8.14.0"
celery = "^5.3.6"
redis = "4.0.1"
coverage = "^7.4.1"
drf-yasg = "^1.21.7"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
